requirements.txt
langchain==1.2.0
langchain-core==1.2.5
langchain-google-genai==4.1.2
streamlit==1.52.2
google-generativeai==0.8.6
python-dotenv==1.2.1
